# src/Shared/TimeForCode.Shared/Api/Authentication/Models/DiscoveryDocumentResponse.cs
using System.ComponentModel.DataAnnotations;
using System.Text.Json.Serialization;

namespace TimeForCode.Shared.Api.Authentication.Models
{
    /// <summary>
    /// Response model for the discovery document endpoint.
    /// </summary>
    public class DiscoveryDocumentResponse
    {
        /// <summary>
        /// Token issuer
        /// </summary>
        [JsonPropertyName("issuer")]
        [Required]
        public required string Issuer { get; init; }

        /// <summary>
        /// JWKS Uri
        /// </summary>
        [JsonPropertyName("jwks_uri")]
        [Required]
        public required string JwksUri { get; init; }
    }
}